pyproject.toml
[project]
name = "fyp-rag"
version = "0.1.0"
description = "FAST-NUCES BS FYP Handbook 2023 - Retrieval-Augmented Generation Assistant"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pymupdf>=1.24.0",
    "langchain-text-splitters>=0.3.0",
    "sentence-transformers>=3.0.0",
    "chromadb>=0.5.0",
    "rank-bm25>=0.2.2",
    "streamlit>=1.38.0",
    "groq>=0.11.0",
    "python-dotenv>=1.0.0",
    "numpy>=1.26.0",
    "tqdm>=4.66.0",
    "nltk>=3.9.0",
    "reportlab>=4.2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/fyp_rag"]

[tool.hatch.metadata]
allow-direct-references = true
